Signal 처리를 별도의 파일로 분리 및 전송여부 필드명 수정
