+++ mailing/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class MailingConfig(AppConfig):
     name = 'mailing'
+
+    def ready(self):
+        import mailing.signal
